moved all the code into one unique .py
